First Twitter Scraper Implementation
The program is now capable of scraping a Tweet or a user profile for information and saving that data to a JSON. It is also capable of saving specific fields that will be useful for sentiment analysis.

Next steps will be to extract lots of tweets relevant to stock market information and compile the data for training.